chore(CI): clean tests
removed duplicated tests
reduce ci window size to imporve render speed
limit frameRate on renderring tests
