+++ test/components/PostEffects_TAA.test.ts
@@ -2,7 +2,9 @@ import { test, expect, end, delay } from '../util'
 import { Camera3D, CameraUtil, Engine3D, Object3D, PostProcessingComponent, Scene3D, TAAPost, View3D } from '@orillusion/core';
 
 await test('Post TAAPost test', async () => {
-    let suc = await Engine3D.init();
+    await Engine3D.init();
+    Engine3D.frameRate = 10;
+
     let view = new View3D();
     view.scene = new Scene3D();
     view.camera = CameraUtil.createCamera3DObject(view.scene, "camera");
